fix(gui-visual-factory): using float for every entrypoint with variadic templates

+++ gui/plugins/entities/Tile/Tile.hpp
@@ -30,7 +30,7 @@ class Tile : public ecs::AEntity {
 };
 
 extern "C" {
-    std::unique_ptr<ecs::IEntity> tileEntrypoint(double x, double y, int type) {
+    std::unique_ptr<ecs::IEntity> tileEntrypoint(float x, float y, int type) {
         return std::make_unique<Tile>(sf::Vector2f(x, y), BiomeTypes_e(type));
     }
 }
